feat(chat): add model selector to quick app modal
